reorganize preprocessing code
