+++ mom6/data_structure/data_preprocessing/batch_attr_mod_cefi_rel_path.py
