+++ mom6/data_structure/data_preprocessing/batch_preprocess_forecast_nwa_raw.json
