+++ mom6/data_structure/data_preprocessing/batch_preprocess_hindcast_nwa_raw_uovo.json
